requirements.txt
annotated-types==0.6.0
anyio==4.3.0
appnope==0.1.4
backcall==0.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cognitojwt==1.4.1
cryptography==41.0.7
decorator==5.1.1
Deprecated==1.2.14
distro==1.9.0
dnspython==2.6.1
ecdsa==0.18.0
email_validator==2.1.1
exceptiongroup==1.2.0
fastapi==0.110.0
filelock==3.13.3
fsspec==2024.3.1
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.22.0
idna==3.6
ipython==7.34.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mpmath==1.3.0
multion==0.3.16
networkx==3.2.1
numpy==1.26.4
oauthlib==3.2.2
openai==1.14.3
orjson==3.9.15
packaging==24.0
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pyasn1==0.5.1
pycparser==2.21
pydantic==2.6.4
pydantic-extra-types==2.6.0
pydantic-settings==2.2.1
pydantic_core==2.16.3
Pygments==2.17.2
pytesseract==0.3.10
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.4.1
rsa==4.9
safetensors==0.4.2
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
sympy==1.12
tokenizers==0.15.2
torch==2.2.1
tqdm==4.66.2
traitlets==5.14.2
transformers==4.39.1
typing_extensions==4.10.0
ujson==5.9.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
wrapt==1.16.0
